requirements.txt
torch
ftfy
regex
tqdm
Pillow
numpy
streamlit
git+https://github.com/openai/CLIP.git
